Move `MapStringifier` to the `org.spine3.base` package.
