+++ client/src/main/java/org/spine3/base/Stringifiers.java
@@ -20,15 +20,10 @@
 
 package org.spine3.base;
 
-import com.google.common.primitives.Ints;
-import com.google.common.primitives.Longs;
-
 import java.lang.reflect.Type;
 import java.util.Map;
-import java.util.regex.Pattern;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.collect.Maps.newHashMap;
 import static org.spine3.base.StringifierRegistry.getStringifier;
 
 /**
@@ -98,131 +93,33 @@ public class Stringifiers {
     }
 
     /**
-     * The stringifier for the {@code Map} classes.
+     * Obtains {@code Stringifier} for map.
      *
-     * <p> The converter for the type of the elements in the map
-     * should be registered in the {@code StringifierRegistry} class
-     * for the correct usage of the {@code MapStringifier} converter.
-     *
-     * @param <K> the type of the keys in the map
-     * @param <V> the type of the values in the map
+     * @param keyClass   the class of keys are maintained by this map
+     * @param valueClass the class  of mapped values
+     * @param <K>        the type of keys are maintained by this map
+     * @param <V>        the type of mapped values
+     * @return the stringifier for the map
      */
-    protected static class MapStringifier<K, V> extends Stringifier<Map<K, V>> {
-
-        private static final char DEFAULT_ELEMENTS_DELIMITER = ',';
-        private static final String ESCAPE_SEQUENCE = "\\";
-        private static final String KEY_VALUE_DELIMITER = ESCAPE_SEQUENCE + ':';
-
-        /**
-         * The delimiter for the passed elements in the {@code String} representation,
-         * {@code DEFAULT_ELEMENTS_DELIMITER} by default.
-         */
-        private final String delimiter;
-        private final Class<K> keyClass;
-        private final Class<V> valueClass;
-
-        public MapStringifier(Class<K> keyClass, Class<V> valueClass) {
-            super();
-            this.keyClass = keyClass;
-            this.valueClass = valueClass;
-            this.delimiter = ESCAPE_SEQUENCE + DEFAULT_ELEMENTS_DELIMITER;
-        }
-
-        /**
-         * That constructor should be used when need to use
-         * a custom delimiter of the elements during conversion.
-         *
-         * @param keyClass   the class of the key elements
-         * @param valueClass the class of the value elements
-         * @param delimiter  the delimiter for the passed elements via string
-         */
-        public MapStringifier(Class<K> keyClass, Class<V> valueClass, String delimiter) {
-            super();
-            this.keyClass = keyClass;
-            this.valueClass = valueClass;
-            this.delimiter = ESCAPE_SEQUENCE + delimiter;
-        }
-
-        @Override
-        protected String toString(Map<K, V> obj) {
-            final String result = obj.toString();
-            return result;
-        }
-
-        @Override
-        protected Map<K, V> fromString(String s) {
-            final String[] buckets = s.split(Pattern.quote(delimiter));
-            final Map<K, V> resultMap = newHashMap();
-
-            for (String bucket : buckets) {
-                saveConvertedBucket(resultMap, bucket);
-            }
-            Ints.stringConverter();
-            return resultMap;
-        }
-
-        private Map<K, V> saveConvertedBucket(Map<K, V> resultMap, String element) {
-            final String[] keyValue = element.split(Pattern.quote(KEY_VALUE_DELIMITER));
-            checkKeyValue(keyValue);
-
-            final String key = keyValue[0];
-            final String value = keyValue[1];
-
-            try {
-                final K convertedKey = convert(keyClass, key);
-                final V convertedValue = convert(valueClass, value);
-                resultMap.put(convertedKey, convertedValue);
-                return resultMap;
-            } catch (Throwable ignored) {
-                throw conversionArgumentException("Occurred exception during the conversion");
-            }
-        }
-
-        @SuppressWarnings("unchecked") // It is OK because class is verified.
-        private static <I> I convert(Class<I> elementClass, String elementToConvert) {
-
-            if (isInteger(elementClass)) {
-                return (I) Ints.stringConverter()
-                               .convert(elementToConvert);
-            }
-
-            if (isLong(elementClass)) {
-                return (I) Longs.stringConverter()
-                                .convert(elementToConvert);
-            }
-
-            if (isString(elementClass)) {
-                return (I) elementToConvert;
-            }
-
-            final I convertedValue = Stringifiers.fromString(elementToConvert, elementClass);
-            return convertedValue;
-
-        }
-
-        private static void checkKeyValue(String[] keyValue) {
-            if (keyValue.length != 2) {
-                final String exMessage =
-                        "Illegal key - value format, key value should be " +
-                        "separated with single `:` character";
-                throw conversionArgumentException(exMessage);
-            }
-        }
-
-        private static IllegalConversionArgumentException conversionArgumentException(String msg) {
-            return new IllegalConversionArgumentException(new ConversionException(msg));
-        }
+    public static <K, V> Stringifier<Map<K, V>> mapStringifier(Class<K> keyClass,
+                                                               Class<V> valueClass) {
+        final Stringifier<Map<K, V>> mapStringifier = new MapStringifier<>(keyClass, valueClass);
+        return mapStringifier;
     }
 
-    private static boolean isString(Class<?> aClass) {
-        return String.class.equals(aClass);
-    }
-
-    private static boolean isLong(Class<?> aClass) {
-        return Long.class.equals(aClass);
-    }
-
-    private static boolean isInteger(Class<?> aClass) {
-        return Integer.class.equals(aClass);
+    /**
+     * @param keyClass   the class of keys are maintained by this map
+     * @param valueClass the class  of mapped values
+     * @param delimiter  the delimiter for the passed map elements via string
+     * @param <K>        the type of keys are maintained by this map
+     * @param <V>        the type of mapped values
+     * @return the stringifier for the map
+     */
+    public static <K, V> Stringifier<Map<K, V>> mapStringifier(Class<K> keyClass,
+                                                               Class<V> valueClass,
+                                                               String delimiter) {
+        final Stringifier<Map<K, V>> mapStringifier =
+                new MapStringifier<>(keyClass, valueClass, delimiter);
+        return mapStringifier;
     }
 }
